moved data processing to domain module

+++ data/local/src/main/java/com/example/local/DailyTemperatureConverter.kt
@@ -20,7 +20,7 @@ class DailyTemperatureConverter {
             val split = it.split(", ")
             val number = split[0].substring(1)
             val description = split[1].subSequence(0, split[1].length - 1)
-            Daily(temp = number.toInt(), description = description.toString())
+            Daily(temp = number.toDouble(), description = description.toString())
         })
     }
 }
