+++ domain/src/main/java/com/example/forecast/domain/model/Hourly.kt
@@ -1,3 +1,3 @@
 package com.example.forecast.domain.model
 
-data class Hourly(val temp: Int, val description: String)
+data class Hourly(val temp: Double, val description: String)
